grn, service list, scheduling, po, audit issues and changes
Service List - Print Error.
GRN - Add Vessel, Evenet, Charge Order NO.
GRN - Add Po - COncat PO with Event and Vessel.
Scheduling. Date Filter Error, Time Error.
New PO - Not Printing.
maintain audit in event module (insert, update, delete, bulk delete)
maintain audit in vessel module (insert, update, delete, bulk delete)"

+++ api/app/Http/Controllers/GRNController.php
@@ -7,10 +7,12 @@ use App\Models\DocumentType;
 use App\Models\GRN;
 use App\Models\GRNDetail;
 use App\Models\Product;
+use App\Models\PurchaseInvoice;
 use App\Models\PurchaseOrder;
 use App\Models\StockLedger;
 use App\Models\Warehouse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Validator;
 
 class GRNController extends Controller
@@ -24,6 +26,9 @@ class GRNController extends Controller
 		$document_identity = $request->input('document_identity', '');
 		$document_date = $request->input('document_date', '');
 		$purchase_order_id = $request->input('purchase_order_id', '');
+		$charge_order_id = $request->input('charge_order_id', '');
+		$event_id = $request->input('event_id', '');
+		$vessel_id = $request->input('vessel_id', '');
 
 		$search = $request->input('search', '');
 		$page =  $request->input('page', 1);
@@ -32,7 +37,10 @@ class GRNController extends Controller
 		$sort_direction = ($request->input('sort_direction') == 'ascend') ? 'asc' : 'desc';
 
 		$data = GRN::LeftJoin('supplier as s', 's.supplier_id', '=', 'good_received_note.supplier_id')
-			->LeftJoin('purchase_order as p', 'p.purchase_order_id', '=', 'good_received_note.purchase_order_id');
+			->LeftJoin('purchase_order as p', 'p.purchase_order_id', '=', 'good_received_note.purchase_order_id')
+			->LeftJoin('charge_order as c', 'c.charge_order_id', '=', 'p.charge_order_id')
+			->LeftJoin('event as e', 'e.event_id', '=', 'c.event_id')
+			->LeftJoin('vessel as v', 'v.vessel_id', '=', 'c.vessel_id');
 		$data = $data->where('good_received_note.company_id', '=', $request->company_id);
 		$data = $data->where('good_received_note.company_branch_id', '=', $request->company_branch_id);
 
@@ -40,18 +48,23 @@ class GRNController extends Controller
 		if (!empty($purchase_order_id)) $data = $data->where('good_received_note.purchase_order_id', '=',  $purchase_order_id);
 		if (!empty($document_identity)) $data = $data->where('good_received_note.document_identity', 'like', '%' . $document_identity . '%');
 		if (!empty($document_date)) $data = $data->where('good_received_note.document_date', '=',  $document_date);
+		if (!empty($charge_order_id)) $data = $data->where('p.charge_order_id', '=',  $charge_order_id);
+		if (!empty($event_id)) $data = $data->where('e.event_id', '=',  $event_id);
+		if (!empty($vessel_id)) $data = $data->where('v.vessel_id', '=',  $vessel_id);
 
 		if (!empty($search)) {
 			$search = strtolower($search);
 			$data = $data->where(function ($query) use ($search) {
 				$query
 					->where('s.name', 'like', '%' . $search . '%')
+					->OrWhere('e.event_code', 'like', '%' . $search . '%')
+					->OrWhere('v.name', 'like', '%' . $search . '%')
 					->OrWhere('p.document_identity', 'like', '%' . $search . '%')
 					->OrWhere('good_received_note.document_identity', 'like', '%' . $search . '%');
 			});
 		}
 
-		$data = $data->select("good_received_note.*", "s.name as supplier_name", "p.document_identity as purchase_order_no");
+		$data = $data->select("good_received_note.*", "s.name as supplier_name", "p.document_identity as purchase_order_no", "e.event_code", "v.name as vessel_name", "c.document_identity as charge_order_no");
 		$data =  $data->orderBy($sort_column, $sort_direction)->paginate($perPage, ['*'], 'page', $page);
 
 		return response()->json($data);
@@ -272,6 +285,23 @@ class GRNController extends Controller
 			return $this->jsonResponse('Permission Denied!', 403, "No Permission");
 		$data  = GRN::where('good_received_note_id', $id)->first();
 		if (!$data) return $this->jsonResponse(['good_received_note_id' => $id], 404, "Good Received Note Not Found!");
+
+
+		$validate = [
+			'main' => [
+				'check' => new GRN,
+				'id' => $id,
+			],
+			'with' => [
+				['model' => new PurchaseInvoice],
+			]
+		];
+
+		$response = $this->checkAndDelete($validate);
+		if ($response['error']) {
+			return $this->jsonResponse($response['msg'], $response['error_code'], "Deletion Failed!");
+		}
+
 		$data->delete();
 		GRNDetail::where('good_received_note_id', $id)->delete();
 		StockLedger::where('document_id', $id)->delete();
@@ -286,6 +316,24 @@ class GRNController extends Controller
 			if (isset($request->good_received_note_ids) && !empty($request->good_received_note_ids) && is_array($request->good_received_note_ids)) {
 				foreach ($request->good_received_note_ids as $good_received_note_id) {
 					$data = GRN::where(['good_received_note_id' => $good_received_note_id])->first();
+
+
+					$validate = [
+						'main' => [
+							'check' => new GRN,
+							'id' => $good_received_note_id,
+						],
+						'with' => [
+							['model' => new PurchaseInvoice],
+						]
+					];
+
+					$response = $this->checkAndDelete($validate);
+					if ($response['error']) {
+						return $this->jsonResponse($response['msg'], $response['error_code'], "Deletion Failed!");
+					}
+
+
 					$data->delete();
 					GRNDetail::where('good_received_note_id', $good_received_note_id)->delete();
 					StockLedger::where('document_id', $good_received_note_id)->delete();
